==> 0_read_convert_data.ipynb <==
import pandas as pd
import seaborn as sns
import json
import matplotlib.pyplot as plt
%matplotlib inline

# Change following path to your telegram data export
path = './DataExport_05_05_2020/result.json'

with open(path, encoding='utf-8', errors='ignore') as json_data:
     data = json.load(json_data, strict=False)

df = pd.DataFrame(data['chats']['list'][0]['messages'][:])

df = df[df["type"]=="message"].copy()

df = df[~df['text'].isna()]

# Removing message from bot
robot_id = df[df['from']=='Guillaume le Robot']['from_id'].unique().astype(int)[0]

# Removing command messages sent to bot
df = df[~df['text'].astype(str).str.contains("bot_command")]
df = df[df['from']!='Guillaume le Robot']

df.to_csv('telegram_data_20200505.csv')

%history
